+++ src/main/java/nextstep/subway/excetpion/CustomExceptionHandler.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line.exception;
+package nextstep.subway.excetpion;
 
+import nextstep.subway.line.exception.LineNotFoundException;
 import nextstep.subway.station.exception.StationNotFoundException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
@@ -7,8 +8,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@RestControllerAdvice("nextstep.subway.line.ui")
-public class LineExceptionHandler {
+@RestControllerAdvice("nextstep.subway")
+public class CustomExceptionHandler {
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity handleException(Exception e) {
